Resolve match before compile (#2734)
#### Changes
A nested match is one where a single case may match on multiple
destructors, organised in a tree. A nested match case may also match on
a constant value. A simple match case may only match on a single
constructor. The process of compiling/denesting/flattening nested
matches, transforms nested matches into `if statements` and simple
matches.

Previously the flattening was done during resolution, but this PR moves
it further back in the pipeline so the resolved AST does not have the
flattening applied to it. The flattening is still computed during
resolution, but it is stored in a new field instead of being applied to
the resolved AST. That way, the IDE can derive code navigation from the
unflattened AST, but still emit warnings produced by the flattening.
Using the flattened results is done in verification and compilation by
traversing the new field.

I think more idiomatically would be not to have the newly added field
`flattened` inside `NestedMatchExpr/Stmt`, and instead fully replace the
`NestedMatchExpr/Stmt` node, since that would prevent having to add code
in the backends. However, we don't have a good mechanism for replacing
AST nodes yet. Also, doing that would require being able to set up a
transformation that runs after resolution, so it doesn't affect the
IDE's code navigation, but before verification and code generation, so
both of those can benefit from it, and we don't have a good mechanism
for that either. I also wonder whether that would require cloning the
AST right after resolution so it's preserved for code navigation.
However, I think such mechanisms could benefit code generators that
produce warnings after doing other transformations, so it's not a ton of
work just for nested matches.

Code related to the flattening of nested matches has been moved out of
`Resolver.cs` and into a new file `MatchFlattener.cs`. Code related to
resolving nested matches has been moved out of resolver and into the
related AST nodes. Simple matches no longer have resolution code since
they're instantiated already resolved. Resolution of nested matches is
currently done in two phases, the 'Check' and 'Resolve' phase, that can
be merged, but I'll leave that for a follow-up PR.

The was a small transformation that was done during match flattening,
which transforms explicit type annotations on bound case variables, by
replacing them with a let statement around the body of the case, where
the bound let variable carries the explicit type annotation. This
transformation has been moved to the resolution phase of nested matches.
Possibly we don't need this transformation, but I leave removing or
moving it for a follow-up PR.

--- 

Fix a bug related to matches and ghost constructors. In the following
program, D0 is non-ghost, but previously the match was still considered
to be ghost because XY has constructors other than D0 that are ghost.
```
  method P0(xy: XY) returns (r: int) {
-    // the following match statement is ghost, because it indirectly mentions a ghost constructor of type XY
+   // the following match statement is non-ghost, because it does not mention a ghost constructor of type XY
    match xy
-    case D0(_) => r := 0; // error: assignment to r in a ghost context
+   case D0(_) => r := 0;
     case any =>
  }
```

### Testing

- Some commented assertions in the test
`DefinitionTests.DatatypesAndMatches` have been uncommented, to test the
additionally supported behavior.
- Update
[Test/dafny0/GhostDatatypeConstructors-Resolution.dfy](https://github.com/dafny-lang/dafny/pull/2734/files#diff-2dee899b94141f686e4dc3684f1a75c9be9da78d6f0d1e0a2daea33735ea89a4)
to capture new behavior.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>