Jar files, both library and executable, are now created as part of build on Java (#3355)
Fixes #2827

Adds the creation of a java jar file to the build step on a Java
platform.
- If there is a Main method, the jar is an executable jar. So a simple
A.dfy can be built as `dafny build -t:java A.dfy`
and then run as `java -jar A.jar`
- If there is no Main entry point, all the generated class files are
assembled into a library Jar file that can be used on the
classpath as a java library.
- In both cases, the DafnyRuntime library is included in the generated
jar
- The .java and .class artifacts are retained if --spill-translation is
true (not the default) or if the legacy CLI is being used (so as not to
break old workflows)
- The jar file is placed in the directory specified by --output (which
is the working directory by default)

Note that the --output option can specifies the _name_ of the output jar
file, as in Q.jar or zzz/Q.jar,
but it is not the name of the directory in which to put the jar file. In
java builds, the build artifacts are placed in
a directory whose location is the path and name of the jar file, without
'-jar' and with '-java'. (This behavior is unchanged)

Having a library or executable jar simplifies the user's task in
figuring out how to use the built artifacts.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

Co-authored-by: davidcok <davidcok@github.com>