fix: Cross-modular override check (#3223)
Fixes 2500.

To achieve this, we first give the override axiom the visibility of the
overriding function. To ensure we only use the axiom in case we pass the
check, we introduce an intermediate visibility level. The ensures clause
of the overridden function can reference the underlying function, so we
need to substitute the overridden function by the overriding function in
the override check.

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>