Separate execution and compilation code of back-ends (#3407)
This is a refactoring that separates translation from Dafny to a string
in a particular language, from other language specific tasks, such as
running a program in that language or building a native executable
artifact for that language.

While I think this refactoring is good on its own, the main driver for
it is to enable another refactoring. I'd like to extract logic that
doesn't relate to string manipulation out of `SinglePassCompiler` by
moving it to a base class, so that this logic can be reused when
compiling Dafny to an AST (as opposed to compiling to a string or
concrete syntax tree). This PR removes the base class from
`SinglePassCompiler`, freeing it up to have another base class to which
logic can be extracted.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>