Avoid missing errors in BitvectorsMore
With the new automated conversions between int/nat and small bit
vectors in shift expressions, more errors are possible, and this
test previously didn’t have a high enough limit to catch them all.
It still passed only because the error ordering hadn’t changed, so
the set under the limit was the same.