Report PropagateFailure/Extract methods as being deprecated (#3444)
Fixes #2693

This PR generates a warning if the `PropagateFailure` or `Extract`
member involved in a supposed failure-compatible type is a method. A
future PR (at or after Dafny version 4) will remove the support.

So, with this PR, the `IsFailure`, `PropagateFailure`, and `Extract`
members must all be functions.

The PR also removes the previous limitation that `IsFailure` or
`PropagateFailure` be ghost. Not only was this check done quite early
during resolution (before other ghost checking), but (more to the point)
it is fine to allow these behaviors (for example, in ghost methods). The
cost is that the errors generated may not be as direct, since they will
be complaining about automatically generated code.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>