feat: Consistent "create_<ctorname>" method for record types (#3034)
Currently in C# and Java, compiled datatypes expose a `create_A(...)`
factory method for each `A` data constructor. However, datatypes with a
single data constructor, also called "record types", will instead expose
a single `create` method. This is slightly nicer for hand-written
external code, but creates an annoying special case for generated code.

This change ensures that record types expose BOTH a `create_<ctorname>`
method for the single data constructor AND a `create` method that thunks
to it, to maintain backwards compatibility but also provide consistent
naming.

Added cases to test this to the existing `Extern.dfy` test case, which
now shows that at least all supported languages in that test case
provide such consistently-named (within a language at least) factory
methods, and don't require direct calls to constructors (note that
Python has built-in support such that for a class `Foo`, `Foo(...)` can
be redefined as needed, and so provides the same flexibility as a
factory method).