fix: dotnet test (or /runAllTests) prints incorrectly with --unicode-char (#3608)
Fixes #3562. Fixes #3595.

The root cause of #3562 is actually an issue with the "halt recovery"
functionality used by `/runAllTests`/`dotnet test` not doing the right
thing in `--unicode-char` mode: in that mode strings are only printed
like strings if the static type of the expression is known to be a
`string`, but the halt recovery AST node doesn't correctly pass on the
knowledge that the captured string at runtime is in fact a string.
Running the `RunAllTestsOption.dfy` test with `--unicode-char` produces
similar looking failures:

```
...
 TestModule.Passing1: PASSED
 Passing1: PASSED
 Passing2: PASSED
 Failing1: FAILED
-	RunAllTestsOption.dfy(21,2): expectation violation
+	['R', 'u', 'n', 'A', 'l', 'l', 'T', 'e', 's', 't', 's', 'O', 'p', 't', 'i', 'o', 'n', '.', 'd', 'f', 'y', '(', '2', '1', ',', '2', ')', ':', ' ', 'e', 'x', 'p', 'e', 'c', 't', 'a', 't', 'i', 'o', 'n', ' ', 'v', 'i', 'o', 'l', 'a', 't', 'i', 'o', 'n']
...
```

Added this tes, fixed the incorrect use of the unresolved
`Type.String()` in the rewriter, and fixed the implementation of
EmitHaltRecovery in some compilers that didn't work with
`--unicode-char`.

Also adding the fix to #3595: `TestAttribute.dfy` was redirecting stdout
instead of stderr as intended, which USUALLY happened to work unless
your `dotnet test` produced different stdout output. The test was
accidentally turned off in #3554 so this change also restores it.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>