Make tests easier in preparation for Z3 4.12.1 (#3613)
Make several of the test easier to prove by adding intermediate
assertions, making things opaque, or adding
`{:vcs_split_on_every_assert}`. This will be necessary for the upgrade
to Z3 4.12.1, and is a portion of #3400, but is something that we should
be able to merge independently and include both before and after the Z3
upgrade.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.