fix: Override checks with `canCall` precondition  (#3479)
We strengthen this check by performing it on the visibility level of the
implementing function, but restricting the enabled axiom to the callers
of the trait. To achieve this, we add the canCall precondition to the
axiom. Additionally, we consider imported trait functions when computing
the visibility levels.

Fixes #2500.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>