Generate Rust-assert warning (#3706)
In Rust, one writes an assertion using a function named `assert!`. I
believe the `!` at the end of the name is a convention for telling the
caller that this function may panic. Thus, to assert that the condition
`x == 0` holds in Rust, one writes

``` rust
assert!(x == 0);
```

To a Dafny programmer, this looks totally wrong, because it seems the
_negation_ of `x == 0` is being asserted. That is, to the eyes of a
Dafny programmer, the Rust function call looks like

``` dafny
assert !(x == 0);
```

If someone used to Rust writes Dafny, there's a chance of accidentally
using the Rust `assert!` and thus ending up with the negation of what's
intended. This PR detects this situation and outputs a warning about it.
More precisely, if a legal program contains `assert` followed by `!`
followed `(`, with no space between the `assert` and the `!`, then a
warning is output.

The warning is not likely to be spurious or be a nuisance for Dafny
programmers, because it is unlikely that a Dafny program would omit this
space on purpose, and even less likely that a Dafny programmer would
have any objections to adding a space there to prevent the warning.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>