chore: Fix release downloads verification workflow (#3678)
Fixes #3712 

These have been broken for several releases. Fixing two issues:

1. The OS identifiers were out of date.
2. Both the workflow steps and quicktest.sh created files like `a.dfy`,
and on some OS' they interfered with each other (probably depending on
how they decided on the current directory). The quick fix was to reorder
them, but do we really need both `quicktest.sh` and what seems like
equivalent workflow steps?

Also fixed the NuGet version, mostly related to Z3 path issues, but also
lots of small bits of rot because of these workflows not being run.

Tested with runs from my fork:
* https://github.com/robin-aws/dafny/actions/runs/4368772189
* https://github.com/robin-aws/dafny/actions/runs/4368722815

Note that one set fails on non-windows machines because `quicktest.sh`
actually changed behavior with the change in string and character
printing in Dafny 4.0. I considered changing the workflow to use the
checked-in quicktest.sh to fix this, but since this is a legitimate
(although small) breakage in what we shipped, it's better to test that
part of the release. Since this PR also fixes `quicktest.sh`, it should
pass for the next release (and hopefully we can verify this with a
nightly release after this is merged).

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.

---------

Co-authored-by: John Tristan <trjohnb@amazon.com>