Automated test coverage analysis (#3486)
Run unit and integration tests with coverage analysis enabled and upload
the results in the test results artifact, both in Cobertura and
marked-up HTML source. Also prints a summary of those coverage results
in the logs.

There's one key improvement I'd like to make, to add the coverage
summary summary as a comment on PRs, as shown
[here](https://github.com/marketplace/actions/code-coverage-summary).
Doing that requires figuring out some permission changes.

To fully view the results, download the `test-coverage-results` artifact
from the `test-coverage-analysis` job and look in the `coverage-html`
directory. Or, if you want to re-generate the HTML, or another coverage
summary format, use the `reportgenerator` tool on the files in
`coverage-cobertura`.

Note that, for some reason, the `DafnyLanguageServer.Test` package only
gets good results when run with `dotnet coverage` as opposed to the
`dotnet test --collect`, and these results are in a different format.
This output is automatically converted to the Cobertura format as part
of the CI job, though, and merged with the rest.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.

---------

Co-authored-by: David Cok <david.r.cok@gmail.com>
Co-authored-by: davidcok <davidcok@github.com>
Co-authored-by: Fabio Madge <fmadge@amazon.com>
Co-authored-by: Remy Willems <rwillems@amazon.com>
Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>
Co-authored-by: stefan-aws <120379523+stefan-aws@users.noreply.github.com>
Co-authored-by: Alex Chew <alex-chew@users.noreply.github.com>