chore: Refactor code to support upcoming new-type-system changes (#3793)
As part of an effort to carve up the large source changes for the
updated Dafny type system, this PR contains various refactorings and
other code improvements that can be reviewed separately.

* When constructing built-in types as `ValuetypeDecl`s, include
type-parameter variance.
* Refactor code that creates the built-in arrow types. (Note, in
separate work, we have noticed errors in the specifications generated
for the `.requires` and `.reads` members of arrow types. This PR does
not attempt to fix those. Instead, this PR just does some refactoring of
the method responsible for creating these.)
* Improve creation of expressions from auto-contract rewriter.
* Add a legacy CLI option `/ntitrace` to spill out debugging information
for the new type inference. (This PR just adds the option. The code that
actually prints debug information will be in a later PR.)
* Refactor some code and extract some methods in order to support calls
in upcoming PRs.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>