feat: Build artifacts ("library" backend and .doo files) (#3913)
Provides a build artifact for outputting after successful verification,
with enough metadata attached to support safe consumption as a library,
as loosely described in this comment:
https://github.com/dafny-lang/dafny/issues/3556#issuecomment-1513871567

Also refactors the functionality used by `dafny audit` to ensure the
library backend rejects `assume` statements as the other backends do,
without having to implement a proper SinglePassCompiler subclass.

Also hides the internal "Simple Dafny" backend, which isn't intended to
be publicly documented, while the new library backend is.

Provides a solution for #3556, but I'm not calling that resolved until
we at least deprecate separate verification features that don't use .doo
files (which isn't in scope for this PR).