Improve resolution to prepare for new type resolution (#3909)
This PR makes various changes and improvements to resolution. The need
for these changes was driven by the upcoming new type inference. The PR
also fixes several issues related to the desugaring of `match`
constructs.

Fixes #3908
Fixes #3915
Fixes #3918
Fixes #3921
Fixes #3922 
Fixes #3923

Changes related to `match`:

* Eliminate `ReplaceTypesWithBoundVariables`. This troublesome method
was changing the AST before desugarings were taking place. This made the
up-and-coming pre-type inference more difficult. And as I was fixing
these other bugs related to `match`, I found that I repeatedly had to
counteract what `ReplaceTypesWithBoundVariables` had done. This because
increasingly more complicated, so I decided to address the real problem
and remove `ReplaceTypesWithBoundVariables`. Amazingly, doing so cause
almost no problems by itself. My conclusion is that this method was here
only for historical purposes, no longer for anything useful.
* Print attributes and types of case patterns
* **fix:** Fix issues with temporary variables introduced during
resolution of case patterns (#3908)
* Introduce `VisitExtendedPattern` method
* Visit patterns during visitation of `NestedMatchStmt`
* Include literal expressions in the `.SubExpressions` of cases
* Don't generate (unused) local/bound variable for literal patterns
* Memoize value computed by `OptimisticallyDesugaredLit`
* Print the RHS literal of symbolic literals in case patterns
* **fix:** Mark ghost status of case variables correctly (#3918)
* **fix:** (**breaking change**) Don't allow downcasts from traits in
case patterns (#3922). These cases previously caused the C# and Java
compiler to emit malformed code. Rather than changing those compilers, I
took the opportunity to disallow such downcasts during resolution. In
the future, we plan to introduce some kind of type-case statement, which
is a better solution for these things, anyway (see also
https://github.com/dafny-lang/dafny/issues/1715). But, technically, this
is a breaking change for anyone who was using the old behavior with the
JavaScript, Go, or Python compilers.
* **fix:** Fix crash that occurred due to malformed desugaring when
encountering `_: int` in pattern (#3923)

Other changes:

* Make working method `CreateResolvedCall` more general to support
parameters and type parameters
* Include `ObjectToDiscard` in the `.SubExpressions` of
`StaticReceiverExpr`
* Generate proper scaffolding around `MemberSelectExpr` and
`FunctionCallExpr` when desugaring iterators
* Refactor some code in the resolution of `AssignOrReturnStmt` to make
routines reusable by pre-resolution
* Move the computation of loop-body surrogate away from the early
name-resolution/type-inference pass. (A consequence of this is that the
warning messages for body-less loops appear in a later phase of
resolution than before.)
* (**breaking change**) Move assumption-variable checks away from the
early name-resolution/type-inference pass. A breaking-change consequence
of this is that type-less assumption variables are no longer inferred to
have type `bool`. (This breaking change is justified as follows:
Assumption variables are rarely, if ever, used. The inference that was
removed only affects assumptions variables that weren't used for
anything, because any other proper use would cause them to be inferred
to be of type `bool`. The breaking-change fix is, at worst, to manually
add the type `: bool` to the declaration of each assumption variable.)
* Make introduction of members of built-in types (like `.Floor` and
`.IsLimit`) more uniform
* **fix:** Fix `:autoReq` to include type parameters of function calls
* **fix:** Re-include the `examples/induction-principle-code/` tests,
which had been "temporarily" excluded for a while. (A recent restriction
that affects those test cases is fixed in this PR.)
* (**breaking change**) Use of `this` when dereferencing static members
is now disallowed in the first phase of object constructors. This makes
the treatment of `this` in that first phase more uniform.
* Change a couple of test cases to no longer rely on cast-free
conversions from a trait to a class. This is in preparation for the new
type inference, which will insist on having such casts explicitly.
* Fix: Perform size checks of bitvector literals in case patterns
(#3915). (At some point in the past, the `git-issue-889b.expect` file of
these tests had been changed, which caused the introduction of this bug
to go unnoticed sooner.)
* **fix:** Check for underspecified type arguments of subset types
(#3921)

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Remy Willems <rwillems@amazon.com>